+++ grid.py
@@ -5,7 +5,7 @@ from colors import Colors
 class Grid:
     def __init__(self):
         self.num_rows = 20
-        self.num_cols = 10
+        self.num_cols = 9
         self.cell_size = 30
         self.grid = [[0 for j in range(self.num_cols)] for i in range(self.num_rows)]
         self.colors = Colors.get_cell_colors()
